# DT-2258: Improve query by alias performance