--- a/src/main/java/org/broadinstitute/consent/http/db/DatasetDAO.java
+++ b/src/main/java/org/broadinstitute/consent/http/db/DatasetDAO.java
@@ -95,14 +95,14 @@ public interface DatasetDAO extends Transactional<DatasetDAO> {
    */
   @UseRowReducer(DatasetReducer.class)
   @SqlQuery("""
-          SELECT d.dataset_id, d.name, d.create_date, d.create_user_id, d.update_date,
-              d.update_user_id, d.object_id, d.dac_id, d.alias, d.data_use, d.translated_data_use,
-              d.dac_approval, d.study_id, d.indexed_date,
-              u.user_id AS u_user_id, u.email AS u_email, u.display_name AS u_display_name,
-              u.create_date AS u_create_date, u.email_preference AS u_email_preference,
-              u.institution_id AS u_institution_id, u.era_commons_id AS u_era_commons_id,
-              k.key, dp.property_value, dp.property_key, dp.property_type, dp.schema_property,
-              dp.property_id
+      SELECT d.dataset_id, d.name, d.create_date, d.create_user_id, d.update_date,
+          d.update_user_id, d.object_id, d.dac_id, d.alias, d.data_use, d.translated_data_use,
+          d.dac_approval, d.study_id, d.indexed_date,
+          u.user_id AS u_user_id, u.email AS u_email, u.display_name AS u_display_name,
+          u.create_date AS u_create_date, u.email_preference AS u_email_preference,
+          u.institution_id AS u_institution_id, u.era_commons_id AS u_era_commons_id,
+          k.key, dp.property_value, dp.property_key, dp.property_type, dp.schema_property,
+          dp.property_id
       FROM dataset d
       LEFT JOIN users u on d.create_user_id = u.user_id
       LEFT JOIN dataset_property dp ON dp.dataset_id = d.dataset_id