--- a/src/main/java/org/broadinstitute/consent/http/resources/TDRResource.java
+++ b/src/main/java/org/broadinstitute/consent/http/resources/TDRResource.java
@@ -40,7 +40,7 @@ public class TDRResource extends Resource {
   public Response getApprovedUsers(@Auth DuosUser duosUser,
       @PathParam("identifier") String identifier) {
     try {
-      Dataset dataset = datasetService.findDatasetByIdentifier(duosUser.getUser(), identifier);
+      Dataset dataset = datasetService.findMinimalDatasetByIdentifier(duosUser.getUser(), identifier);
       if (Objects.isNull(dataset)) {
         throw new NotFoundException("Could not find dataset " + identifier);
       }
@@ -60,7 +60,7 @@ public class TDRResource extends Resource {
   public Response getDatasetByIdentifier(@Auth DuosUser duosUser,
       @PathParam("identifier") String identifier) {
     try {
-      Dataset dataset = datasetService.findDatasetByIdentifier(duosUser.getUser(), identifier);
+      Dataset dataset = datasetService.findMinimalDatasetByIdentifier(duosUser.getUser(), identifier);
       if (Objects.isNull(dataset)) {
         throw new NotFoundException("Could not find dataset " + identifier);
       }