### Addresses
https://broadworkbench.atlassian.net/browse/DT-2258

### Summary
This PR makes two primary changes:
1. Minimize the data returned by the following endpoint:
  - GET /api/tdr/{identifier}
2. Streamline the query for approved users by calling the minimized dataset call:
  - GET /api/dataset/{identifier}/approvedUsers
  - GET /api/tdr/{identifier}/approvedUsers

Some prod vs local timings that all use the same dataset identifier:
| API | Prod | This PR | % Improvement |
| -------- | ------- | ------- | ------- |
| GET /api/tdr/{identifier} | 6599 ms | 449 ms | ~93% |
| GET /api/dataset/{identifier}/approvedUsers | 6675 ms | 530 ms | ~92% |
| GET /api/tdr/{identifier}/approvedUsers | 6600 ms | 730 ms | ~89% |


----
Have you read [CONTRIBUTING.md](../CONTRIBUTING.md) lately? If not, do that first.

- Label PR with a Jira ticket number and include a link to the ticket
- Label PR with a security risk modifier [no, low, medium, high]
- PR describes scope of changes
- Get a minimum of one thumbs worth of review, preferably two if enough team members are available
- Get PO sign-off for all non-trivial UI or workflow changes
- Verify all tests go green
- Test this change deployed correctly and works on dev environment after deployment